Add "location" tracking for rooms
One of the things that's hard about using the ptg page is that you
don't know where things are located. This allows you to set the
location on a room, which is displayed in a 3rd column, and will be
persistent after a new set of now (thus not requiring everyone to
remember to include it in their updates).

Change-Id: I12bac56e29cc5974d5c925f23eb94945e6e36421