Sort room list
This provides a sorting of the room list by iterating the total list
of rooms, and displaying information per room if available. Without
this the order of rooms jumps on each save based on the fact that it's
a dict/object and has no natural sort order. That can feel very
confusing.

Change-Id: I5375374091296cd7c23a3a8e2272aefcf6b2d896