Also find room for tracks in additional rooms
The get_track_room function failed to also look in
additional rooms to find the room for a given track.

If the event is not running in the current day, we
should default to the first day of the event (useful
for tests and the day before the event starts).

Change-Id: I143eb3b490b3536679e0608064209ec71ed37499