Auto-hyperlink URLs
If someone issues a command like

    #upgrades now Upgrades SIG https://etherpad.openstack.org/p/ffu-ptg-rocky

then make the URL a hyperlink so that viewers of the webpage can
easily jump to the etherpad to learn about the topic being discussed
and make a quick but informed decision whether it's worth them going
to that room.

To avoid duplicating the word iteration logic, we do this by extending
the existing "hashtag" helper; however this widens its scope beyond
merely formatting hashtags, so also rename the helper to reflect this.

Change-Id: I8e21ab9a1966a09a7b4ec1742eeb578cfd0a3f49