Split long lines
Use textwrap to they are split on word boundaries.  400 characters
is used as an approximation of a safe line length without dealing
with the complexity of determining the exact value for specific
conditions.  In practice, we've seen lines truncated at 435 chars.

We sleep between issuing chunks, however, we sleep less than the
normal flood protection interval so the bot still appears responsive.

Change-Id: I87d26228f0e025d911b29a1662a069a2b6d0e497