Add more usage hints
Since there's already a hint on how to tell the bot to book a room,
and some people have been slightly struggling with updating track
info, add a couple of hints making it more obvious how to do this
without having to consult a larger README.rst.  This should hopefully
encourage more people to help keep track info up to date.

Also include a link to the nicely rendered version of README.rst on
GitHub, so that if people are looking for more detailed help, they can
easily find it.

Change-Id: I3596d69dc84b635bc3a67ff7223a7ccfad27308e