ptgbot should still ask for identify-map cap
In the recent SSL/SASL introduction, we removed the
on_welcome() and on_cap() functions which were thought
to be taken care of by the introduced mixins.

Turns out it was not the case and the bot needs the
identify-msg cap in order to function, so we should restore
that code.

Change-Id: I11f94ddc2be7e3ff6592dd2c2e5fbfb43f6309f7