Handle all schedule in a single table
The bookable rooms functionality was bolted on top of the "normal"
schedule table using a separate table in the JSON and the config.

Now that bookable rooms are shown together with "regular" rooms,
we can merge the two concept and pay back that technical debt.

This is an incompatible JSON change so the database will have to
be ~reload-ed from configuration.

Change-Id: Ia37b6202b363614418b562b693b3dd807621aa97