Show check-ins per room in "Currently at the PTG" page
Include tooltips on the "Currently at the PTG" page, so that hovering
the mouse over a particular track will show IRC nicks for everyone
currently checked into that track.  If noone has checked in to the
track yet, give a hint about how to do it.

Thanks a lot to Jay Bryant for giving me the idea for this after we
bumped into each other on the plane to Denver ;-)

Change-Id: I1b86c1f403981b066ad096548a1754fadd6b72ad