Recognise when user forgets # prefix for a track check-in
Since the in / #in command allows free-form check-ins when the
location is not prefixed with '#', this would allow people to
check into locations like 'nova' which would not be recognised
as an official track location.  As a result their attendance
would not show up in the tooltips on the schedule web page.

So automatically detect when the user is checking into a track
but forgot to use the '#' prefix, and add it automatically for
them.  This DWIM behaviour should avoid some confusion.

Also when matching / matched against known tracks, the lower-cased
version will be used.  This assumes that all registered tracks are
lower-case.

Change-Id: Ic88ba926ba9e3187da6046972344bc2b31e3c4f1