Display count of attendees in each room on web page
Currently it's not obvious that you can see who's checked into a room
by hovering the mouse over one of the room badges. This in turn makes
it less likely for people to check in, since they either won't be
aware of the feature, or will forget about it.

So game-ify check-ins by adding a new column to the "Current
discussion topics" table which displays the number of check-ins within
that room.

Story: 2005657
Task: 30945
Change-Id: I8d2bc716d39c70ff096f17b4da9ccf0a26bfc34f