No longer display location
The location code was created to allow for emergency relocation of
meetings, however in a virtual setting that is tracked using the "url"
feature. When location and urls differ it can create a lot of confusion.

It's better to stop displaying location for virtual PTGs. As an
emergency fix, let's stop displaying the location.

Change-Id: Ibb9d96e1fa7f686e31cc8d9b51e3826af8c4b015