Turn MOTD into a list of messages
The message of the day used to be a single message with an importance
level. It's actually useful to be able to have several distinct
messages and add/remove them separately.

In preparation for that change, turn the motd entry in the json DB into
an array of messages. Gracefully migrate previous format DB into the
new one. Commands stay unchanged for now and just manipulate the first
entry.

Change-Id: Id98281dba0c6b3cffe1a359d6e0c250b99084bb1